+++ examples/event-source-mapping/versions.tf
@@ -2,6 +2,8 @@ terraform {
   required_version = ">= 0.12.26"
 
   required_providers {
-    aws = ">= 3.19"
+    aws    = ">= 3.27"
+    random = ">= 2"
   }
 }
+