+++ modules/alias/main.tf
@@ -11,7 +11,7 @@ data "aws_lambda_alias" "existing" {
 }
 
 resource "aws_lambda_alias" "no_refresh" {
-  count = var.create && ! var.use_existing_alias && ! var.refresh_alias ? 1 : 0
+  count = var.create && !var.use_existing_alias && !var.refresh_alias ? 1 : 0
 
   name        = var.name
   description = var.description
@@ -29,7 +29,7 @@ resource "aws_lambda_alias" "no_refresh" {
 }
 
 resource "aws_lambda_alias" "with_refresh" {
-  count = var.create && ! var.use_existing_alias && var.refresh_alias ? 1 : 0
+  count = var.create && !var.use_existing_alias && var.refresh_alias ? 1 : 0
 
   name        = var.name
   description = var.description
@@ -90,7 +90,7 @@ resource "aws_lambda_permission" "version_triggers" {
   statement_id       = lookup(each.value, "statement_id", each.key)
   action             = lookup(each.value, "action", "lambda:InvokeFunction")
   principal          = lookup(each.value, "principal", format("%s.amazonaws.com", lookup(each.value, "service", "")))
-  source_arn         = lookup(each.value, "source_arn", lookup(each.value, "service", null) == "apigateway" ? "${lookup(each.value, "arn", "")}/*/*/*" : null)
+  source_arn         = lookup(each.value, "source_arn", null)
   source_account     = lookup(each.value, "source_account", null)
   event_source_token = lookup(each.value, "event_source_token", null)
 }
@@ -104,7 +104,7 @@ resource "aws_lambda_permission" "qualified_alias_triggers" {
   statement_id       = lookup(each.value, "statement_id", each.key)
   action             = lookup(each.value, "action", "lambda:InvokeFunction")
   principal          = lookup(each.value, "principal", format("%s.amazonaws.com", lookup(each.value, "service", "")))
-  source_arn         = lookup(each.value, "source_arn", lookup(each.value, "service", null) == "apigateway" ? "${lookup(each.value, "arn", "")}/*/*/*" : null)
+  source_arn         = lookup(each.value, "source_arn", null)
   source_account     = lookup(each.value, "source_account", null)
   event_source_token = lookup(each.value, "event_source_token", null)
 }