+++ variables.tf
@@ -86,7 +86,7 @@ variable "kms_key_arn" {
 }
 
 variable "memory_size" {
-  description = "Amount of memory in MB your Lambda Function can use at runtime. Valid value between 128 MB to 3008 MB, in 64 MB increments."
+  description = "Amount of memory in MB your Lambda Function can use at runtime. Valid value between 128 MB to 10,240 MB (10 GB), in 64 MB increments."
   type        = number
   default     = 128
 }
@@ -278,6 +278,16 @@ variable "create_unqualified_alias_allowed_triggers" {
 variable "allowed_triggers" {
   description = "Map of allowed triggers to create Lambda permissions"
   type        = map(any)
+  default     = {}
+}
+
+############################################
+# Lambda Event Source Mapping
+############################################
+
+variable "event_source_mapping" {
+  description = "Map of event source mapping"
+  type        = any
   default     = {}
 }